Разбивает main.js на модули и настраивает сборку
