+++ src/components/more-button.js
@@ -1,0 +1,7 @@
+const createButtonMoreTemplate = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {createButtonMoreTemplate};
