+++ src/components/movies-quantity.js
@@ -1,0 +1,9 @@
+const createMoviesQuantityTemplate = () => {
+  return (
+    `<section class="footer__statistics">
+      <p>130 291 movies inside</p>
+    </section>`
+  );
+};
+
+export {createMoviesQuantityTemplate};
